% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{as.theta.vcov}
\alias{as.theta.vcov}
\title{Get theta parameterisation of a covariance structure}
\usage{
as.theta.vcov(Sigma, corrs.only = FALSE)
}
\arguments{
\item{Sigma}{a covariance matrix}

\item{corrs.only}{return only values corresponding to the correlation matrix parameters?}
}
\value{
the corresponding \code{theta} parameter vector
}
\description{
Get theta parameterisation of a covariance structure
}
