% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{bf_matrix}
\alias{bf_matrix}
\title{Function to evaluate basis functions (created via \code{make_basis()}) at point locations provided and return them in a matrix}
\usage{
bf_matrix(
  basis.functions,
  point.locations,
  bf.matrix.type = c("sparse", "dense")
)
}
\arguments{
\item{basis.functions}{an object created by \code{make_basis()}}

\item{point.locations}{either a matrix or data.frame describing the point locations at which the basis functions are to be evaluated.}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to create sparse or dense matrix.}
}
\value{
a matrix (sparse or dense depending on \code{bf.matrix.type})
}
\description{
Calculates a matrix of basis function values at the point locations supplied (rows are locations and columns are basis functions). Primarily used internally for model fitting, also used to predict the basis at new locations.
}
\examples{
data("gorillas")
# Create basis function nodes on the locations of presence records and quadrature
bfs <- make_basis(k = 100, data = gorillas)
# Calculate the basis function matrix at the point events in gorillas
X <- bf_matrix(bfs, gorillas[gorillas$pres == 1, c("x", "y")])
}
