% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{biplot.glmmTMB}
\alias{biplot.glmmTMB}
\title{Create a biplot from a multivarite LGCP}
\usage{
\method{biplot}{glmmTMB}(
  x,
  ...,
  alpha = 0.5,
  load.names,
  score.col,
  load.col,
  load.name.cex = 1,
  show.top.n,
  vmax.rotation = FALSE,
  rm.bias = T
)
}
\arguments{
\item{x}{a fitted model object from \code{mvlgcp}.}

\item{...}{additional parameters as in \code{stats::biplot}}

\item{alpha}{a numeric in [0,1] indicating the scaling of arrows/lines representing factor loading lengths. Defaults to 0.5}

\item{load.names}{(optional) a character vector of length \code{object$n_reponses} of labels to be used on response-specific loadings. Default is \code{NULL} which means the fitted model's \code{response.id} will be used to determine labels.}

\item{score.col}{(optional) a vector of length \code{object$n_basis} of colours to be used for scores.}

\item{load.col}{(optional) a vector of length \code{object$n_reponses} of colours to be used on response-specific loadings.}

\item{load.name.cex}{(optional) a numeric indicating the scaling of response label size.}

\item{show.top.n}{(optional) a numeric in [0,\code{object$n_reponses}] indicating a subset of the response-specific loadings (from largest to smallest).}

\item{vmax.rotation}{a logical indicating whether to perform a varimax rotation of the loadings matrix prior to plotting.}

\item{rm.bias}{a logical indicating whether to remove bias from both the scores and factor loadings by scaling them #NEEDS WORK}
}
\value{
See \code{stats::biplot}
}
\description{
This function creates a biplot of basis function coefficients (scores) forming the \code{n_factors} shared latent fields approximating the multivarate latent fields in a LGCP. Arrows/lines represent response-specific factor loadings.
}
\examples{
dat <- lansing[lansing$tree \%in\% c("blackoak", "hickory", "maple"), ]
dat$tree <- factor(as.character(dat$tree)) # remove unused tree species

# set up the basis functions with dimension k = 100
bfs <- make_basis(k = 100, dat)

# fit a multivariate LGCP to the point patterns of "blackoak", "hickory" and "maple" trees
m <- mvlgcp(pt ~ (1 | tree), data = dat, weights = dat$wt, basis.functions = bfs, response.id = dat$tree)

biplot(m)
}
