% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{isLMM.glmmTMB}
\alias{isLMM.glmmTMB}
\alias{refit.glmmTMB}
\title{support methods for parametric bootstrapping}
\usage{
isLMM.glmmTMB(object)

refit.glmmTMB(object, newresp, ...)
}
\arguments{
\item{object}{a fitted glmmTMB object}

\item{newresp}{a new response vector}

\item{...}{other}
}
\description{
see \code{\link[lme4]{refit}} and \code{\link[lme4]{isLMM}} for details
}
