% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{compois}
\alias{compois}
\title{Conway-Maxwell Poisson distribution parameterized by exact mean}
\usage{
compois(link = "log")
}
\arguments{
\item{link}{(character) link function. Only "log" is impemented.}
}
\description{
Conway-Maxwell Poisson distribution parameterized by exact mean
}
