% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{.adjustX}
\alias{.adjustX}
\title{Adjust a model matrix
When not rank deficient, do nothing.
When rank deficient matrix, drop columns.}
\usage{
.adjustX(X, tol = NULL, why_dropped = FALSE)
}
\arguments{
\item{X}{model matrix}

\item{tol}{non-negative tolerance for testing for "practically zero" singular values (passed to Matrix::rankMatrix())}

\item{why_dropped}{logical indicating whether or not to provide information about sets of collinear predictors (not yet implemented)}
}
\description{
Adjust a model matrix
When not rank deficient, do nothing.
When rank deficient matrix, drop columns.
}
\keyword{internal}
