% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{.checkRankX}
\alias{.checkRankX}
\title{check for identifiability of fixed effects matrices X, Xzi, Xd}
\usage{
.checkRankX(data.tmb, rank_check = c("warn", "adjust", "stop", "skip"))
}
\description{
check for identifiability of fixed effects matrices X, Xzi, Xd
}
\keyword{internal}
