% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop.special2}
\alias{drop.special2}
\title{drop terms matching a particular value from an expression}
\usage{
drop.special2(x, value = quote(offset), preserve = NULL)
}
\arguments{
\item{x}{formula}

\item{value}{term to remove from formula}

\item{preserve}{(integer) retain the specified occurrence of "value"}
}
\description{
drop terms matching a particular value from an expression
}
\keyword{internal}
