% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropHead}
\alias{dropHead}
\title{return a formula/expression with a given value stripped, where
it occurs as the head of a term}
\usage{
dropHead(term, value)
}
\description{
return a formula/expression with a given value stripped, where
it occurs as the head of a term
}
\examples{
dropHead(~a+offset(b),quote(offset))
dropHead(~a+poly(x+z,3)+offset(b),quote(offset))
}
\keyword{internal}
