% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtruncated_nbinom2}
\alias{dtruncated_nbinom2}
\alias{dtruncated_poisson}
\alias{dtruncated_nbinom1}
\title{truncated distributions}
\usage{
dtruncated_nbinom2(x, size, mu, k = 0, log = FALSE)

dtruncated_poisson(x, lambda, k = 0, log = FALSE)

dtruncated_nbinom1(x, phi, mu, k = 0, log = FALSE)
}
\arguments{
\item{x}{value}

\item{size}{number of trials/overdispersion parameter}

\item{mu}{mean parameter}

\item{k}{truncation parameter}

\item{log}{(logical) return log-probability?}

\item{lambda}{mean parameter}

\item{phi}{overdispersion parameter}
}
\description{
Probability functions for k-truncated Poisson and negative binomial distributions.
}
