% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformulas.R
\name{expandGrpVar}
\alias{expandGrpVar}
\title{apply}
\usage{
expandGrpVar(f)
}
\arguments{
\item{f}{a language object (an atom of a formula)
expandGrpVar(quote(x*y))
expandGrpVar(quote(x/y))}
}
\description{
apply
}
