% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractForm}
\alias{extractForm}
\title{extract terms with a given head from an expression/formula}
\usage{
extractForm(term, value)
}
\arguments{
\item{term}{expression/formula}

\item{value}{head of terms to extract}
}
\value{
a list of expressions
}
\description{
extract terms with a given head from an expression/formula
}
\examples{
extractForm(~a+offset(b),quote(offset))
extractForm(~c,quote(offset))
extractForm(~a+offset(b)+offset(c),quote(offset))
}
\keyword{internal}
