% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{family_params}
\alias{family_params}
\title{Retrieve family-specific parameters}
\usage{
family_params(object)
}
\arguments{
\item{object}{glmmTMB object}
}
\value{
a named numeric vector
}
\description{
Most conditional distributions have only parameters governing their location
(retrieved via \code{predict}) and scale (\code{sigma}). A few (e.g. Tweedie, Student t, ordered beta)
are characterized by one or more additional parameters.
}
