% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbx}
\alias{fbx}
\title{(f)ind (b)ars e(x)tended: recursive}
\usage{
fbx(term, debug = FALSE, specials = character(0), default.special = "us")
}
\arguments{
\item{term}{a formula or piece of a formula}

\item{debug}{(logical) debug?}

\item{specials}{list of special terms}

\item{default.special}{character: special to use for parenthesized terms - i.e. random effects terms with unspecified structure
1. atom (not a call or an expression): NULL
2. special, i.e. foo(...) where "foo" is in specials: return term
3. parenthesized term: \emph{if} the head of the head is | (i.e.
   it is of the form (xx|gg), then convert it to the default
   special type; we won't allow pathological cases like
   ((xx|gg)) ... [can we detect them?]}
}
\description{
(f)ind (b)ars e(x)tended: recursive
}
\keyword{internal}
