% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{genpois}
\alias{genpois}
\title{Generalized Poisson distribution}
\usage{
genpois(link = "log")
}
\arguments{
\item{link}{(character) link function. Only "log" is impemented.}
}
\description{
Generalized Poisson distribution
}
