% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{get_bfs_coeff}
\alias{get_bfs_coeff}
\title{Extract the basis function coefficients from a model fitted via \code{mvlgcp()}}
\usage{
get_bfs_coeff(object)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} model object.}
}
\value{
a matrix with k (the number/dimension of the basis functions used to approx. the latent field) rows and d (= 1 in the univariate case, or the number of latent fields for a multivariate LGCP) columns.
}
\description{
This function extracts the random coefficient terms (at the mode of the Laplace approximation) from a fitted model. It will automatically account for a reduced rank approach when the fitted model is a multivariate LGCP.
}
\examples{
# using a subset of the Lansing Wood data:
dat <- lansing[lansing$tree == "blackoak", ]

# set up the basis functions with dimension k = 100
bfs <- make_basis(k = 100, dat)

# fit univariate LGCP to the blackoak point pattern data
m <- mvlgcp(pt ~ 1, data = dat, weights = dat$wt, basis.functions = bfs)

# extract the field
b <- get_bfs_coeff(m)
plot(b)
}
