% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{get_loadings}
\alias{get_loadings}
\title{Extract the factor loadings from a model fitted via \code{mvlgcp()}}
\usage{
get_loadings(object)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} model object.}
}
\value{
a matrix with m (the number of responses) rows and d (the number of latent fields for a multivariate LGCP) columns.
}
\description{
This function sets up the appropriate basis function matrix for the \code{newdata} provided, then multiplies by the random coefficient terms (at the mode of the Laplace approximation).
}
\examples{
dat <- lansing[lansing$tree \%in\% c("blackoak", "hickory", "maple"), ]
dat$tree <- factor(as.character(dat$tree)) # remove unused tree species

# set up the basis functions with dimension k = 100
bfs <- make_basis(k = 100, dat)

# fit a multivariate LGCP to the point patterns of "blackoak", "hickory" and "maple" trees
m <- mvlgcp(pt ~ (1 | tree), data = dat, weights = dat$wt, basis.functions = bfs, response.id = dat$tree)

# extract the factor loadings
Lambda <- get_loadings(m)
plot(Lambda)
}
