% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{getCapabilities}
\alias{getCapabilities}
\title{List model options that glmmTMB knows about}
\usage{
getCapabilities(what = "all")
}
\arguments{
\item{what}{(character) which type of model structure to report on
("all","family","link","covstruct")}
}
\description{
List model options that glmmTMB knows about
}
\note{
these are all the options that are \emph{defined} internally; they have not necessarily all been \emph{implemented} (FIXME!)
}

