% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getME.glmmTMB}
\alias{getME.glmmTMB}
\alias{getME}
\title{Extract or Get Generalize Components from a Fitted Mixed Effects Model}
\usage{
\method{getME}{glmmTMB}(
  object,
  name = c("X", "Xzi", "Z", "Zzi", "Xdisp", "theta", "beta", "b", "Gp", "cnms", "flist"),
  ...
)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} object}

\item{name}{of the component to be retrieved}

\item{\dots}{ignored, for method compatibility}
}
\description{
Extract or Get Generalize Components from a Fitted Mixed Effects Model
}
\note{
The \code{"cnms"} and \code{"flist"} values are specific to the
conditional component of the model (i.e., they do not include
zero-inflation or dispersion terms). Users can extract the analogous
values for those components via \code{object$modelInfo$reTrms[[component]]$cnms}
(or \code{...$flist}), where \code{component} is \code{"zi"} or \code{"disp"}.
See \code{\link[lme4]{getME}} for definitions of individual components.
}
\seealso{
\code{\link[lme4]{getME}}
}
