% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{terms.glmmTMB}
\alias{terms.glmmTMB}
\alias{model.matrix.glmmTMB}
\title{Methods for extracting developer-level information from \code{glmmTMB} models}
\usage{
\method{terms}{glmmTMB}(x, component = "cond", part = "fixed", ...)

\method{model.matrix}{glmmTMB}(object, component = "cond", part = "fixed", ...)
}
\arguments{
\item{x}{a fitted \code{glmmTMB} object}

\item{component}{model component ("cond", "zi", or "disp"; not all models contain all components)}

\item{part}{whether to return results for the fixed or random effect part of the model (at present only \code{part="fixed"} is implemented for most methods)}

\item{\dots}{additional arguments (ignored or passed to \code{\link{model.frame}})}

\item{object}{a fitted \code{glmmTMB} object}
}
\description{
Methods for extracting developer-level information from \code{glmmTMB} models
}
