% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_load}
\alias{gt_load}
\title{Load data from system file, updating glmmTMB objects}
\usage{
gt_load(fn, verbose = FALSE, mustWork = FALSE)
}
\arguments{
\item{fn}{partial path to system file (e.g. test_data/foo.rda)}

\item{verbose}{print names of updated objects?}

\item{mustWork}{fail if file not found?}
}
\description{
Load data from system file, updating glmmTMB objects
}
