% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inForm}
\alias{inForm}
\title{test formula: does it contain a particular element?}
\usage{
inForm(form, value)
}
\description{
test formula: does it contain a particular element?
}
\examples{
inForm(z~.,quote(.))
inForm(z~y,quote(.))
inForm(z~a+b+c,quote(c))
inForm(z~a+b+(d+e),quote(c))
f <- ~ a + offset(x)
f2 <- z ~ a
inForm(f,quote(offset))
inForm(f2,quote(offset))
}
\keyword{internal}
