% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\docType{data}
\name{lansing}
\alias{lansing}
\title{Lansing Woods Point Pattern (as a data.frame)}
\format{
An object of class \code{"data.frame"} with 2251 rows and 3 variables:
\describe{
  \item{x}{Horizontal coordinate.}
  \item{y}{Vertical coordinate.}
  \item{tree}{Botanical classification of trees: hickories; maples; red, white and black oaks; and miscellaneous.}
  \item{pt}{Binary indicating whether the row is a point event (1) or a quadrature point (0).}
  \item{wt}{The weighting for each quadrature point. Set to 1e-10 at the point events.}
 }
}
\usage{
data(lansing)
}
\description{
Locations and botanical classification of trees in Lansing Woods. The data come from an investigation of a 924 ft x 924 ft (19.6 acre) plot in Lansing Woods, Clinton County, Michigan USA by D.J. Gerrard. The data give the locations of 2251 trees and their botanical classification (into hickories, maples, red oaks, white oaks, black oaks and miscellaneous trees). The original plot size (924 x 924 feet) has been rescaled to the unit square. This dataset comes from the \code{spatstat} package, convert to a \code{data.frame} here for use in \code{glmmTMB::mvlgcp()}. Note that a regular grid of quadrature points has been added (for each species), as needed for the multivariate model. See \code{spatstat.data::lansing}.
}
\references{
Gerrard, D.J. (1969) Competition quotient: a new measure of the competition affecting individual forest trees. Research Bulletin 20, Agricultural Experiment Station, Michigan State University.

Turner, R. and Baddeley, A. (2005), SPATSTAT: an R package for analyzing spatial point patterns, \emph{Journal of Statistical Software}, \bold{12}.
}
\keyword{datasets}
