% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.glmmTMB}
\alias{logLik.glmmTMB}
\title{Extract the log likelihood of a glmmTMB model}
\usage{
\method{logLik}{glmmTMB}(object, ...)
}
\value{
object of class \code{logLik} with attributes
\item{val}{log likelihood}
\item{nobs,nall}{number of non NA observations initially supplied to TMB}
\item{df}{number of parameters}
}
\description{
Extract the log likelihood of a glmmTMB model
}

