% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{make_basis}
\alias{make_basis}
\title{Create 2D basis functions for use in \code{glmmTMB::mvlgcp()}}
\usage{
make_basis(
  k,
  data,
  from.package = c("scampr", "mgcv", "FRK"),
  coord.names = c("x", "y"),
  longlat = FALSE,
  raw = FALSE,
  ...
)
}
\arguments{
\item{k}{integer describing the desired basis dimension.}

\item{data}{a data frame containing the two coordinates described by \code{coord.names}.}

\item{from.package}{a character string that describing the package from which to create basis functions.}

\item{coord.names}{vector of character strings describing the names of the coordinates in 'data'. Ordered horizontal then vertical axes}

\item{longlat}{specific to \code{scampr} basis functions, a logical indicating whether the coordinates are in Longitude and Latitude so that geodesic distances are used (the radius is also extended to the maximum geodesic distance between nodes). Defaults to \code{FALSE}.}

\item{raw}{specific to \code{mgcv} basis functions, a logical of whether to return raw or transformed values (mgcv does the latter by default, assists fitting algorithms to converge). Defaults to \code{FALSE}.}

\item{...}{additional arguments to be added for the underlying basis construction. See \code{mgcv::smoothCon()} and \code{FRK::auto_basis()} for example.}
}
\value{
an object of class specific to the basis function type required. To be passed to bf_matrix()
}
\description{
Uses data provided to place points across the range of the provided coordinates. Options include basis functions as in the packages \code{"scampr"} (see Dovers et al. 2023), \code{"mgcv"} (see Wood 2017), or \code{"FRK"} (see Zammit-Mangion and Cressie 2021). Note that the final basis dimension may not match \code{k}, particularly on unequal coordinate ranges - trial-and-error may be needed to get your desired basis dimension.
}
\examples{
# Create basis function nodes on the locations of presence records and quadrature
bfs <- make_basis(k = 100, data = gorillas)
}
\references{
Wood, S. N. (2017), Generalized additive models: an introduction with R, \emph{CRC Press}.

Zammit-Mangion, A. and Cressie, N. (2021), FRK: An R package for spatial and spatio-temporal prediction with large datasets, \emph{Journal of Statistical Software}, \bold{98}, 1–48.

Dovers, E., Brooks, W., Popovic, G. C., and Warton, D. I. (2023), Fast, Approximate Maximum Likelihood Estimation of Log-Gaussian Cox Processes, \emph{Journal of Computational and Graphical Statistics}, \bold{32}, 1660–1670.
}
