% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{print.VarCorr.glmmTMB}
\alias{print.VarCorr.glmmTMB}
\title{Printing The Variance and Correlation Parameters of a \code{glmmTMB}}
\usage{
\method{print}{VarCorr.glmmTMB}(
  x,
  digits = max(3, getOption("digits") - 2),
  comp = "Std.Dev.",
  formatter = format,
  maxdim = 10,
  ...
)
}
\arguments{
\item{x}{a result of \code{\link{VarCorr}(<glmmTMB>)}.}

\item{digits}{number of significant digits to use.}

\item{comp}{a string specifying the component to format and print.}

\item{formatter}{a \code{\link{function}}.}

\item{maxdim}{maximum number of SDs/correlations to print}

\item{...}{optional further arguments, passed to \code{\link[reformulas]{formatVC}}}
}
\description{
Printing The Variance and Correlation Parameters of a \code{glmmTMB}
}
