% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{prune_basis}
\alias{prune_basis}
\title{Prune basis functions (created via \code{make_basis()}) that do not intersect the data supplied}
\usage{
prune_basis(basis.functions, data, n_non_zero = 1, prune = TRUE)
}
\arguments{
\item{basis.functions}{an object created by \code{make_basis()}}

\item{data}{a data frame containing the coordinates associated with the supplied \code{basis.functions}}

\item{n_non_zero}{an integer describing the number of non-zero points required for pruning. Default is 1.}

\item{prune}{a logcial indicating whether to prune the basis functions, or simply identify problem ones. Default is \code{TRUE}.}
}
\value{
a basis functions object as in \code{make_basis} (or a vector of indices describing problem basis functions if \code{prune=FALSE})
}
\description{
Computes the basis function matrix and checks for columns that do not have at least \code{n_non_zero} non-zero entires. If \code{prune=TRUE} basis functions not meeting the requirement will be removed.
}
\examples{
data("gorillas")
# Create basis function nodes on the locations of presence records and quadrature
bfs <- make_basis(k = 100, data = gorillas)
# prune the basis functions according to a subset of the gorillas data
new.bfs <- prune(bfs, gorillas[1:1000, ])
}
