% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{refit}
\alias{isGLMM}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lme4}{\code{\link[lme4:isGLMM]{isGLMM()}}, \code{\link[lme4:refit]{refit()}}}
}}

