% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinstalling}
\alias{reinstalling}
\title{Reinstalling binary dependencies}
\description{
The \code{glmmTMB} package depends on several upstream packages, which it
uses in a way that depends heavily on their internal (binary) structure.
Sometimes, therefore, installing an update to one of these packages will
require that you re-install a \emph{binary-compatible} version of \code{glmmTMB},
i.e. a version that has been compiled with the updated version of the upstream
package.

While you'll always get a warning when there is a mismatch between
the version(s) of the upstream package(s) used to build \code{glmmTMB} and
the version(s) you have installed, these are only warnings. Such mismatches
may cause problems, but the problems are likely to be obvious (crashes); if
\code{glmmTMB} seems to be working OK anyway, you can ignore the warning.

If not:
\itemize{
\item You may be able to install a freshly built binary (of the development version) from \code{glmmTMB}'s
\href{https://glmmtmb.r-universe.dev/glmmTMB}{R-universe page} via
\code{install.packages('glmmTMB', repos = c('https://glmmtmb.r-universe.dev', 'https://cloud.r-project.org'))}
\item If you have development tools (compilers etc.) installed, you
should be able to re-install a binary-compatible version of the package by running
\code{install.packages("glmmTMB", type="source")}. If you want to install
the development version of \code{glmmTMB} instead, you can use
\code{remotes::install_github("glmmTMB/glmmTMB/glmmTMB")}.
(On Windows, you can install development tools following the instructions at
\url{https://cran.r-project.org/bin/windows/Rtools/}; on MacOS, see
\url{https://mac.r-project.org/tools/}.)
}

If R-universe isn't working for you,
you don't have development tools installed, and you can't
install them for some reason(and so can't install packages with compiled code from source),
please file an issue (with full details about your
operating system and R version) asking the maintainers to build and
post an appropriate binary version of the package (we may or may not
be able to help, depending on your system and how much time we have).
}
