% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_fns.R
\name{rrTerms}
\alias{rrTerms}
\title{Extract  reduced-rank (rr) factor loadings and latent variables}
\usage{
rrTerms(x)
}
\arguments{
\item{x}{a fitted \code{glmmTMB} model}
}
\value{
A named list with three components (\code{cond}, \code{zi}, \code{disp}). 
  Each component contains a list with one element per random effect term. For rr 
  structures, each element contains:
\itemize{
  \item \code{lv} matrix of latent variables with columns named \code{lv1}, \code{lv2},..
  \item \code{fl} matrix of factor loadings
}  
  Non-rr terms return \code{NULL}.
}
\description{
Extracts the factor loadings and latent variables from reduced-rank 
random effect structures in a fitted \code{glmmTMB} model.
}
\examples{
\dontrun{
  # Fit model with reduced-rank structure
  m1_rr <- glmmTMB(abund ~ Species + rr(Species + 0|id, d = 1), data = spider_long)
  m1_rr_trms <- rrTerms(m1_rr)
  m1_rr_trms$cond[[1]]$lv   # latent variables
  m1_rr_trms$cond[[1]]$fl   # factor loadings
}
}
