% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_simcodes}
\alias{set_simcodes}
\title{helper function to modify simulation settings for random effects}
\usage{
set_simcodes(g, val = "zero", terms = "ALL")
}
\arguments{
\item{g}{a TMB object}

\item{val}{a legal setting for sim codes ("zero", "random", or "fix")}

\item{terms}{which terms to apply this to}
}
\description{
This modifies the TMB object \emph{in place} (beware!)
Ultimately this will allow \code{terms} to be a vector of term names,
with a matching \code{val} vector to specify the behaviour for each term
}
