% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_new}
\alias{simulate_new}
\title{Simulate from covariate/metadata in the absence of a real data set (EXPERIMENTAL)}
\usage{
simulate_new(form, data, pars, ..., show_pars = FALSE)
}
\arguments{
\item{form}{a model formula}

\item{data}{a data frame containing all variables listed in the formula,
\emph{including} the response variable (which needs to fall within
the domain of the conditional distribution, and should probably not
be all zeros, but whose value is otherwise irrelevant)}

\item{pars}{a list of parameters containing sub-vectors
(\code{beta}, \code{betazi}, \code{betad}, \code{theta}, etc.) to
be used in the model}

\item{...}{other arguments to \code{glmmTMB} (e.g. \code{family})}

\item{show_pars}{(logical) print structure of parameter vector and stop without simulating?}
}
\description{
Simulate from covariate/metadata in the absence of a real data set (EXPERIMENTAL)
}
