% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spider_long}
\alias{spider_long}
\title{Spider data from CANOCO, long format}
\format{
An object of class \code{data.frame} with 336 rows and 9 columns.
}
\usage{
spider_long
}
\description{
data from spider2 directory, CANOCO FORTRAN package, with trait
variables added; taken from the \code{mvabund} package and converted to long form. Variables:
\itemize{
\item soil.dry
\item bare.sand
\item fallen.leaves
\item moss
\item herb.layer
\item reflection 
\item id
\item Species
\item abund
}
}
\references{
\itemize{
  \item ter Braak, C. J. F. and Smilauer, P. (1998) CANOCO reference manual and user's guide to CANOCO for Windows: software for canonical community ordination (version 4). Microcomputer Power, New York, New York, USA
  \item van der Aart, P. J. M., and Smeenk-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycosidae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology 25, 1-45.
}
}
\keyword{data}
