% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{stripReTrms}
\alias{stripReTrms}
\title{select only desired pieces from results of getXReTrms}
\usage{
stripReTrms(xrt, whichel = c("cnms", "flist"))
}
\description{
select only desired pieces from results of getXReTrms
}

