% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{summary.glmmTMB}
\alias{summary.glmmTMB}
\title{summary for glmmTMB fits}
\usage{
\method{summary}{glmmTMB}(
  object,
  sandwich = FALSE,
  ddf = c("asymptotic", "kenward-roger", "satterthwaite"),
  cluster = getGroups(object),
  ...
)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} object}

\item{sandwich}{use the sandwich estimator for the variance-covariance matrix? (this only works for ML fits, but not for REML fits)}

\item{ddf}{denominator degrees-of-freedom calculation. Default "asymptotic" gives standard Z-statistics
(i.e., 'infinite' denominator df); \code{"kenward-roger"} uses the Kenward-Roger approximation
(see \code{\link{dof_KR}}), which requires a REML fit (an error is thrown otherwise) and a family
with an estimated dispersion parameter (an error is thrown for families such as \code{binomial} or
\code{poisson} that lack one); \code{"satterthwaite"} uses a Satterthwaite approximation, with no such
restrictions. For families other than \code{gaussian}, both approximations are allowed but emit a
warning, because their performance (and theoretical justification) for GLMMs is poorly understood}

\item{cluster}{grouping factor for the sandwich estimator, only used if \code{sandwich==TRUE}.}

\item{...}{unused, for method compatibility}
}
\description{
summary for glmmTMB fits
}
