% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.glmmTMB}
\alias{vcov.glmmTMB}
\title{Extracts the variance covariance structure}
\usage{
\method{vcov}{glmmTMB}(object, full = FALSE, ...)
}
\description{
Extracts the variance covariance structure
}

