% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcovHC.glmmTMB}
\alias{vcovHC.glmmTMB}
\title{Cluster Robust Variance-Covariance Matrix Estimator}
\usage{
\method{vcovHC}{glmmTMB}(x, type = "HC0", sandwich = TRUE, ...)
}
\arguments{
\item{x}{a \code{glmmTMB} object fitted with ML (REML is not supported).}

\item{type}{only "HC0" is supported for \code{glmmTMB} models.}

\item{sandwich}{logical; if \code{TRUE}, return the sandwich estimator,
otherwise only the meat matrix is returned.}

\item{...}{additional arguments passed to \code{\link{meatHC}} and
\code{\link{sandwich}}, in particular `full` and `cluster` arguments.}
}
\value{
A square matrix representing the cluster-robust variance-covariance matrix.
}
\description{
This method for \code{\link[sandwich]{vcovHC}} computes the cluster-robust 
variance-covariance matrix for a \code{glmmTMB} model fitted with ML.
}
\examples{
m <- glmmTMB(count ~ mined + (1 | spp), data = Salamanders, family = nbinom1)
vcov(m)
vcovHC(m)
vcovHC(m, full = TRUE)
}
