% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{vec2im}
\alias{vec2im}
\title{Function to convert a vector (with corresponding coordinates) to spatstat image.}
\usage{
vec2im(vec, x.loc, y.loc, return.matrix = FALSE)
}
\arguments{
\item{vec}{a vector of field values of equal length to locations x.loc and y.loc}

\item{x.loc}{a vector of horizontal locations in 2D}

\item{y.loc}{a vector of vertical locations in 2D}

\item{return.matrix}{a logical indicating whether to return the vector in a matrix of y.loc (rows) and x.loc (columns) instead of the default \code{im} object. Useful if other image plotting is preferred.}
}
\value{
a spatstat image (im) object (or matrix if \code{return.matrix = TRUE})
}
\description{
This function converts the supplied vector into a matrix of vertical (rows) and horizontal (columns) points according to the supplied x, y locations to create a pixel image. Default is to create a \code{spatstat} image object.
}
\examples{
# Get the domain from the gorillas data
domain <- gorillas[gorillas$pres == 0, ]
elevation <- vec2im(vec = domain$elevation, x.loc = domain$x, y.loc = domain$y)
plot(elevation)
}
