% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgcp.R
\name{biplot.glmmTMB}
\alias{biplot.glmmTMB}
\title{Create a biplot from a multivarite LGCP}
\usage{
\method{biplot}{glmmTMB}(
  x,
  ...,
  alpha,
  load.names,
  score.col,
  load.col,
  load.name.cex = 1,
  show.top.n,
  show.responses,
  load.lab.offset,
  do.svd = FALSE,
  incl.ind.flds = TRUE,
  show.all.arrows = FALSE,
  rotate.by.theta
)
}
\arguments{
\item{x}{a fitted model object from \code{mvlgcp}.}

\item{...}{additional parameters as in \code{stats::biplot}}

\item{alpha}{a numeric in [0,1] indicating the scaling of arrows/lines representing factor loading lengths. Defaults to the ratio of the maximum L2 norms for the scores to factor loadings.}

\item{load.names}{(optional) a character vector of length \code{object$n_reponses} of labels to be used on response-specific loadings. Default is \code{NULL} which means the fitted model's \code{response.id} will be used to determine labels.}

\item{score.col}{(optional) a vector of length \code{object$n_basis} of colours to be used for scores.}

\item{load.col}{(optional) a vector of length \code{object$n_reponses} of colours to be used on response-specific loadings.}

\item{load.name.cex}{(optional) a numeric indicating the scaling of response label size.}

\item{show.top.n}{(optional) a numeric in [1,\code{object$n_reponses}] indicating the number of response-specific loadings to show (from largest to smallest).}

\item{show.responses}{(optional) a numeric vector with elements that are response indices (\emph{i.e.} in [1,\code{object$n_reponses}]) indicating a subset of the response-specific loadings to be plotted.}

\item{load.lab.offset}{a numeric, indicating the offset for response labels on the factor loadings. Defaults to 0.1}

\item{do.svd}{a logical indicating whether to additionally perform a Singular Value Decomposition on the approximate basis function coefficient matrix. Default is \code{FALSE}. Primarily used to include the additional independent fields so that the loadings are transformed to represent correlation between point patterns.}

\item{incl.ind.flds}{a logical indicating whether to include basis function coefficients from independent fields (if the model has these) within the Singular Value Decomposition - will only apply when \code{do.svd = TRUE}, in which case the default is \code{TRUE}.}

\item{show.all.arrows}{a logical indicating whether to plot all arrows/lines, even when \code{show.top.n} is used.}

\item{rotate.by.theta}{a numeric, indicating a rotation of the biplot (in Radians).}
}
\value{
See \code{stats::biplot}
}
\description{
This function creates a biplot of basis function coefficients (scores) forming the \code{n_factors} shared latent fields approximating the multivarate latent fields in a LGCP. Arrows/lines represent response-specific factor loadings.
}
\examples{
dat <- lansing[lansing$tree \%in\% c("blackoak", "hickory", "maple"), ]
dat$tree <- factor(as.character(dat$tree)) # remove unused tree species

# set up the basis functions with dimension k = 100
bfs <- make_basis(k = 100, dat)

# fit a multivariate LGCP to the point patterns of "blackoak", "hickory" and "maple" trees
m <- mvlgcp(pt ~ (1 | tree), data = dat, weights = dat$wt, basis.functions = bfs, response.id = dat$tree)

biplot(m)
}
