% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denom_df.R
\name{dof_KR}
\alias{dof_KR}
\title{compute Kenward-Roger degrees of freedom}
\usage{
dof_KR(model)
}
\arguments{
\item{model}{a fitted \code{glmmTMB} object}
}
\value{
a named vector of ddf for each conditional fixed-effect parameter, with attributes 'vcov'
(Kenward-Roger adjusted covariance matrix) and 'se' (the corresponding standard errors)
}
\description{
This function uses an adaptation of the machinery from the \code{pbkrtest} package
to compute the Kenward-Roger approximation of the 'denominator degrees of freedom' for
each fixed-effect coefficient in the conditional model.
}
\details{
this function \emph{should not be used} for models fitted with ML rather than REML;
the theory is only well understood, and the model is only tested, for LMMs (\code{family = "gaussian"}).
Use at your own risk for GLMMs!
}
