% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{.checkRankX}
\alias{.checkRankX}
\title{Check for identifiability of fixed effects matrices X, Xzi, Xd.
When rank_check='adjust', drop columns in X and remove associated parameters.}
\usage{
.checkRankX(TMBStruc, rank_check = c("warning", "adjust", "stop", "skip"))
}
\description{
Check for identifiability of fixed effects matrices X, Xzi, Xd.
When rank_check='adjust', drop columns in X and remove associated parameters.
}
\keyword{internal}
