% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{format_sdvar}
\alias{format_sdvar}
\alias{format_corr}
\title{format columns corresponding to std. dev. and/or variance}
\usage{
format_sdvar(
  reStdDev,
  use.c = "Std.Dev.",
  formatter = format,
  maxlen = 10,
  digits = max(3, getOption("digits") - 2),
  ...
)

format_corr(x, maxdim = Inf, digits = 2, maxlen = 10, ...)
}
\arguments{
\item{reStdDev}{a vector of standard deviations}

\item{formatter}{formatting function}

\item{maxlen}{maximum number of rows to display}

\item{digits}{digits for format}

\item{...}{additional arguments to formatter}

\item{x}{a square numeric matrix)}

\item{maxdim}{maximum number of rows/columns to display}

\item{a}{character vector indicating which scales to include}
}
\description{
format columns corresponding to std. dev. and/or variance

format correlation matrix
}
\examples{
gt_load("test_data/models.rda")
format_sdvar(reStdDev = 1:3, use.c = c("Variance", "Std.Dev."))
format_sdvar(attr(VarCorr(fm1)$cond$Subject, "stddev"))
gt_load("test_data/models.rda")
format_corr(attr(VarCorr(fm1)$cond$Subject, "correlation"))
}
