% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{getME.glmmTMB}
\alias{getME.glmmTMB}
\title{Extract or Get Generalize Components from a Fitted Mixed Effects Model
Method borrowed from lme4}
\usage{
\method{getME}{glmmTMB}(object, name = c("X", "Xzi", "Z", "Zzi", "Xd",
  "theta"), ...)
}
\description{
Extract or Get Generalize Components from a Fitted Mixed Effects Model
Method borrowed from lme4
}

