% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{getReStruc}
\alias{getReStruc}
\title{Calculate random effect structure
calculates number of random effects, number of parameters,
blocksize and number of blocks.}
\usage{
getReStruc(reTrms, ss)
}
\arguments{
\item{reTrms}{random-effects terms list}

\item{ss}{}
}
\value{
a list
\item{blockNumTheta}{number of variance covariance parameters per term}
\item{blockSize}{size (dimension) of one block}
\item{blockReps}{number of times the blocks are repeated (levels)}
\item{covCode}{structure code}
}
\description{
Calculate random effect structure
calculates number of random effects, number of parameters,
blocksize and number of blocks.
}
\examples{
data(sleepstudy, package="lme4")
rt <- lme4::lFormula(Reaction~Days+(1|Subject)+(0+Days|Subject),
                    sleepstudy)$reTrms
rt2 <- lme4::lFormula(Reaction~Days+(Days|Subject),
                    sleepstudy)$reTrms
getReStruc(rt)
}

