% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{getXReTrms}
\alias{getXReTrms}
\title{Create X and random effect terms from formula}
\usage{
getXReTrms(formula, mf, fr, ranOK = TRUE, type = "")
}
\arguments{
\item{formula}{current formula, containing both fixed & random effects}

\item{mf}{matched call}

\item{fr}{full model frame}

\item{ranOK}{random effects allowed here?}

\item{type}{label for model type}
}
\value{
a list composed of
\item{X}{design matrix for fixed effects}
\item{Z}{design matrix for random effects}
\item{reTrms}{output from \code{\link{mkReTrms}} from \pkg{lme4}}
}
\description{
Create X and random effect terms from formula
}

