% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{glmmTMB}
\alias{glmmTMB}
\title{main TMB function}
\usage{
glmmTMB(formula, data = NULL, family = gaussian(), ziformula = ~0,
  dispformula = ~1, weights = NULL, offset = NULL, se = FALSE,
  debug = FALSE)
}
\arguments{
\item{formula}{combined fixed and random effects formula, following lme4 syntac}

\item{data}{data frame}

\item{family}{\code{\link{family}}}

\item{ziformula}{combined fixed and random effects formula for zero-inflation: the default \code{~0} specifies no zero-inflation}

\item{dispformula}{combined fixed and random effects formula for dispersion: the default \code{~0} specifies no zero-inflation}

\item{weights}{}

\item{offset}{}

\item{se}{whether to return standard errors}

\item{debug}{whether to return the preprocessed data and parameter objects, without fitting the model}
}
\description{
main TMB function
}
\examples{
data(sleepstudy, package="lme4")
glmmTMB(Reaction ~ Days + (1|Subject), sleepstudy, debug=TRUE)
glmmTMB(Reaction ~ Days + us(1|Subject), sleepstudy, debug=TRUE)
glmmTMB(Reaction ~ Days + diag(1|Subject), sleepstudy, debug=TRUE)
}

