% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{glmmTMB}
\alias{glmmTMB}
\title{Fit models with TMB}
\usage{
glmmTMB(formula, data = NULL, family = gaussian(), ziformula = ~0,
  dispformula = ~1, weights = NULL, offset = NULL, se = TRUE,
  verbose = FALSE, debug = FALSE)
}
\arguments{
\item{formula}{combined fixed and random effects formula, following lme4
syntax}

\item{data}{data frame}

\item{family}{family (variance/link function) information; see \code{\link{family}} for
details.  As in \code{\link{glm}}, \code{family} can be specified as (1) a character string
referencing an existing family-construction function (e.g. \sQuote{"binomial"}); (2) a symbol referencing
such a function (\sQuote{binomial}); or (3) the output of such a function (\sQuote{binomial()}).
In addition, for families such as \code{betabinomial} that are special to \code{glmmTMB}, family
can be specified as (4) a list comprising the name of the distribution and the link function
(\sQuote{list(family="binomial",link="logit")}).}

\item{ziformula}{a \emph{one-sided} (i.e., no response variable) formula for
    zero-inflation combining fixed and random effects:
the default \code{~0} specifies no zero-inflation.
Specifying \code{~.} will set the right-hand side of the zero-inflation
formula identical to the right-hand side of the main (conditional effects)
formula; terms can also be added or subtracted. \strong{Offset terms
will automatically be dropped from the conditional effects formula.}
The zero-inflation model uses a logit link.}

\item{dispformula}{a \emph{one-sided} formula for dispersion containing only fixed effects: the
default \code{~1} specifies the standard dispersion given any family.
The argument is ignored for families that do not have a dispersion parameter.
For an explanation of the dispersion parameter for each family, see (\code{\link{sigma}}).
The dispersion model uses a log link. 
In Gaussian mixed models, \code{dispformula=~0} fixes the paramameter to be 0, forcing variance into the random effects.}

\item{weights}{weights, as in \code{glm}. Only implemented for binomial and betabinomial families.}

\item{offset}{offset}

\item{se}{whether to return standard errors}

\item{verbose}{logical indicating if some progress indication should be printed to the console.}

\item{debug}{whether to return the preprocessed data and parameter objects,
without fitting the model}
}
\description{
Fit models with TMB
}
\details{
\itemize{
\item binomial models with more than one trial (i.e., not binary/Bernoulli)
must be specified in the form \code{prob ~ ..., weights = N} rather than in
the more typical two-column matrix (\code{cbind(successes,failures)~...}) form.
\item in all cases \code{glmmTMB} returns maximum likelihood estimates - random effects variance-covariance matrices are not REML (so use \code{REML=FALSE} when comparing with \code{lme4::lmer}), and residual standard deviations (\code{\link{sigma}}) are not bias-corrected. Because the \code{\link{df.residual}} method for \code{glmmTMB} currently counts the dispersion parameter, one would need to multiply by \code{sqrt(nobs(fit)/(1+df.residual(fit)))} when comparing with \code{lm} ...
}
}
\examples{
data(sleepstudy, package="lme4")
sleepstudy <- transform(sleepstudy, DaysFac = factor(Days))
fm1 <- glmmTMB(Reaction ~ Days +     (1|Subject), sleepstudy)
fm2 <- glmmTMB(Reaction ~ Days + us  (1|Subject), sleepstudy)
fm3 <- glmmTMB(Reaction ~ Days + diag(1|Subject), sleepstudy)
fm4 <- glmmTMB(Reaction ~ ar1(DaysFac+0|Subject), sleepstudy)
}

