% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatvalues.R
\name{hatvalues.glmmTMB}
\alias{hatvalues.glmmTMB}
\title{Hat values (and other influence measures) for glmmTMB objects}
\usage{
\method{hatvalues}{glmmTMB}(model, diag = TRUE, ...)
}
\arguments{
\item{model}{a \code{glmmTMB} model}

\item{diag}{Get diagonal only?}

\item{\dots}{unused: for method compatibility only}
}
\description{
Hat values (and other influence measures) for glmmTMB objects
}
\examples{
fm0 <- glmmTMB(weight ~ Time, data=ChickWeight)
fm1 <- lm(weight ~ Time, data = ChickWeight)
all.equal(hatvalues(fm0), unname(hatvalues(fm1)))
fm2 <- glmmTMB(weight ~ Time + diag(Time | Chick), data=ChickWeight)
fm3 <- lme4::lmer(weight ~ Time + (Time || Chick), data = ChickWeight, REML = FALSE)
all.equal(hatvalues(fm2), unname(hatvalues(fm3)), tolerance = 1e-2)
}
