% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{map.theta.propto}
\alias{map.theta.propto}
\title{Set map values for theta to be fixed (NA) for propto}
\usage{
map.theta.propto(ReStruc, map)
}
\arguments{
\item{ReStruc}{a covariance matrix}

\item{corrs.only}{return only values corresponding to the correlation matrix parameters?}
}
\value{
the corresponding \code{theta} parameter vector
}
\description{
Set map values for theta to be fixed (NA) for propto
}
