% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{mkTMBStruc}
\alias{mkTMBStruc}
\title{Extract info from formulas, reTrms, etc., format for TMB}
\usage{
mkTMBStruc(formula, ziformula, dispformula, combForm, mf, fr, yobs, respCol,
  offset, weights, size = NULL, family, se = NULL, call = NULL,
  verbose = NULL, ziPredictCode = "corrected", doPredict = 0,
  whichPredict = integer(0))
}
\arguments{
\item{formula}{conditional formula}

\item{ziformula}{zero-inflation formula}

\item{dispformula}{dispersion formula}

\item{combForm}{combined formula}

\item{mf}{call to model frame}

\item{fr}{frame}

\item{yobs}{observed y}

\item{respCol}{response column}

\item{offset}{offset}

\item{weights}{weights}

\item{size}{number of trials in binomial and betabinomial families}

\item{family}{family object}

\item{se}{(logical) compute standard error?}

\item{call}{original \code{glmmTMB} call}

\item{verbose}{verbosity setting from original \code{glmmTMB} call}

\item{ziPredictCode}{zero-inflation code}

\item{doPredict}{flag to enable sds of predictions}

\item{whichPredict}{which observations in model frame represent predictions}
}
\description{
Extract info from formulas, reTrms, etc., format for TMB
}
\keyword{internal}
