% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{mkTMBStruc}
\alias{mkTMBStruc}
\title{Extract info from formulas, reTrms, etc., format for TMB}
\usage{
mkTMBStruc(formula, ziformula, dispformula, mf, fr, yobs, offset, weights,
  family, link, ziPredictCode = "corrected", doPredict = 0,
  whichPredict = integer(0))
}
\arguments{
\item{formula}{conditional formula}

\item{ziformula}{zero-inflation formula}

\item{dispformula}{dispersion formula}

\item{mf}{call to model frame}

\item{fr}{frame}

\item{yobs}{observed y}

\item{offset}{offset}

\item{weights}{weights}

\item{family}{character}

\item{link}{character}

\item{ziPredictCode}{zero-inflation code}

\item{doPredict}{flag to enable sds of predictions}

\item{whichPredict}{which observations in model frame represent predictions}
}
\description{
Extract info from formulas, reTrms, etc., format for TMB
}
\keyword{internal}

