% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmmTMB}
\alias{predict.glmmTMB}
\title{prediction}
\usage{
\method{predict}{glmmTMB}(object, newdata = NULL, debug = FALSE, re.form,
  allow.new.levels = FALSE, zitype = c("response", "conditional", "zprob"),
  ...)
}
\arguments{
\item{object}{a \code{glmmTMB} object}

\item{newdata}{new data for prediction}

\item{zitype}{for zero-inflated models,
return expected value ("response": (mu*(1-p))),
the mean of the conditional distribution ("conditional": mu),
or the probability of a structural zero ("zprob")?
mean of}
}
\description{
prediction
}
\examples{
data(sleepstudy,package="lme4")
g0 <- glmmTMB(Reaction~Days+(Days|Subject),sleepstudy)
glmmTMB:::predict.glmmTMB(g0,sleepstudy)
}

