% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{print.VarCorr.glmmTMB}
\alias{print.VarCorr.glmmTMB}
\title{Printing The Variance and Correlation Parameters of a \code{glmmTMB}}
\usage{
\method{print}{VarCorr.glmmTMB}(x, digits = max(3, getOption("digits") - 2),
  comp = "Std.Dev.", formatter = format, ...)
}
\arguments{
\item{x}{a result of \code{\link{VarCorr}(<glmmTMB>)}.}

\item{digits}{number of significant digits to use.}

\item{comp}{a string specifying the component to format and print.}

\item{formatter}{a \code{\link{function}}.}

\item{...}{optional further arguments, passed the next \code{\link{print}} method.}
}
\description{
Printing The Variance and Correlation Parameters of a \code{glmmTMB}
}
