% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.glmmTMB}
\alias{residuals.glmmTMB}
\title{Compute residuals for a glmmTMB object}
\usage{
\method{residuals}{glmmTMB}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{a \dQuote{glmmTMB} object}

\item{type}{(character) residual type}

\item{\dots}{ignored, for method compatibility}
}
\description{
Compute residuals for a glmmTMB object
}

