% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{sigma.glmmTMB}
\alias{sigma}
\alias{sigma.glmmTMB}
\title{Extract Residual Standard Deviation or dispersion parameter}
\usage{
\method{sigma}{glmmTMB}(object, ...)
}
\arguments{
\item{object}{a \dQuote{glmmTMB} fitted object}

\item{\dots}{(ignored; for method compatibility)}
}
\description{
For Gaussian models, retrieves the value of the residual
standard deviation; for other model types, retrieves the
dispersion parameter, \emph{however it is defined for that
particular family}.
}

