% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{summary.glmmTMB}
\alias{summary.glmmTMB}
\title{summary for glmmTMB fits}
\usage{
\method{summary}{glmmTMB}(object, ddf = Inf, ...)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} object}

\item{ddf}{denominator degrees-of-freedom calculation (\code{NULL} gives standard Z-statistics
(i.e., 'infinite' denominator df). \code{"KR"} gives Kenward-Roger approximation, which will
be ignored for non-REML fits and is entirely untested for GLMMs (see \code{\link{dof_KR}}).}

\item{...}{unused, for method compatibility}
}
\description{
summary for glmmTMB fits
}
