% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{useRTMB}
\alias{useRTMB}
\title{switch to using RTMB}
\usage{
useRTMB(flag = NULL)
}
\arguments{
\item{flag}{FALSE (default) to use TMB backend; TRUE to use RTMB backend}
}
\description{
switch to using RTMB
}
