% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{VarCorr.glmmTMB}
\alias{VarCorr}
\alias{VarCorr.glmmTMB}
\title{Extract variance and correlation components}
\usage{
\method{VarCorr}{glmmTMB}(x, sigma = 1, rdig = 3)
}
\arguments{
\item{x}{a fitted \code{glmmTMB} model}

\item{sigma}{residual standard deviation (usually set automatically from internal information)}

\item{rdig}{ignored: for \code{nlme} compatibility}
}
\description{
Extract variance and correlation components
}
\keyword{internal}

